Create models module
