+++ ledger_bot/errors/__init__.py
@@ -1,0 +1,5 @@
+"""Additional exceptions."""
+
+from . import airtable_error
+
+AirTableError = airtable_error.AirTableError
