+++ ledger_bot/models/__init__.py
@@ -1,0 +1,8 @@
+"""The data models for ledger_bot."""
+
+from . import bot_message, member, model, transaction
+
+Model = model.Model
+Member = member.Member
+Transaction = transaction.Transaction
+BotMessage = bot_message.BotMessage
